% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{AnnotateImageRequest}
\alias{AnnotateImageRequest}
\title{AnnotateImageRequest Object}
\usage{
AnnotateImageRequest(image = NULL, features = NULL, imageContext = NULL)
}
\arguments{
\item{image}{The image to be processed}

\item{features}{Requested features}

\item{imageContext}{Additional context that may accompany the image}
}
\value{
AnnotateImageRequest object
}
\description{
AnnotateImageRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request for performing Google Cloud Vision API tasks over a user-providedimage, with user-requested features.
}
