% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{AnnotateImageResponse}
\alias{AnnotateImageResponse}
\title{AnnotateImageResponse Object}
\usage{
AnnotateImageResponse(cropHintsAnnotation = NULL, webDetection = NULL,
  labelAnnotations = NULL, safeSearchAnnotation = NULL, error = NULL,
  fullTextAnnotation = NULL, landmarkAnnotations = NULL,
  textAnnotations = NULL, imagePropertiesAnnotation = NULL,
  faceAnnotations = NULL, logoAnnotations = NULL)
}
\arguments{
\item{cropHintsAnnotation}{If present, crop hints have completed successfully}

\item{webDetection}{If present, web detection has completed successfully}

\item{labelAnnotations}{If present, label detection has completed successfully}

\item{safeSearchAnnotation}{If present, safe-search annotation has completed successfully}

\item{error}{If set, represents the error message for the operation}

\item{fullTextAnnotation}{If present, text (OCR) detection or document (OCR) text detection has}

\item{landmarkAnnotations}{If present, landmark detection has completed successfully}

\item{textAnnotations}{If present, text (OCR) detection has completed successfully}

\item{imagePropertiesAnnotation}{If present, image properties were extracted successfully}

\item{faceAnnotations}{If present, face detection has completed successfully}

\item{logoAnnotations}{If present, logo detection has completed successfully}
}
\value{
AnnotateImageResponse object
}
\description{
AnnotateImageResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response to an image annotation request.
}
