% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{BatchAnnotateImagesRequest}
\alias{BatchAnnotateImagesRequest}
\title{Google Cloud Vision API Objects 
Integrates Google Vision features, including image labeling, face, logo, and landmark detection, optical character recognition (OCR), and detection of explicit content, into applications.}
\usage{
BatchAnnotateImagesRequest(requests = NULL)
}
\arguments{
\item{requests}{Individual image annotation requests for this batch}
}
\value{
BatchAnnotateImagesRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:21:48
filename: /Users/mark/dev/R/autoGoogleAPI/googlevisionv1.auto/R/vision_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
BatchAnnotateImagesRequest Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Multiple image annotation requests are batched into a single service call.
}
\seealso{
Other BatchAnnotateImagesRequest functions: \code{\link{images.annotate}}
}
