% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{BatchAnnotateImagesResponse}
\alias{BatchAnnotateImagesResponse}
\title{BatchAnnotateImagesResponse Object}
\usage{
BatchAnnotateImagesResponse(responses = NULL)
}
\arguments{
\item{responses}{Individual responses to image annotation requests within the batch}
}
\value{
BatchAnnotateImagesResponse object
}
\description{
BatchAnnotateImagesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response to a batch image annotation request.
}

