% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{Block}
\alias{Block}
\title{Block Object}
\usage{
Block(property = NULL, blockType = NULL, boundingBox = NULL,
  paragraphs = NULL)
}
\arguments{
\item{property}{Additional information detected for the block}

\item{blockType}{Detected block type (text, image etc) for this block}

\item{boundingBox}{The bounding box for the block}

\item{paragraphs}{List of paragraphs in this block (if this blocks is of type text)}
}
\value{
Block object
}
\description{
Block Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Logical element on the page.
}
