% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{BoundingPoly}
\alias{BoundingPoly}
\title{BoundingPoly Object}
\usage{
BoundingPoly(vertices = NULL)
}
\arguments{
\item{vertices}{The bounding polygon vertices}
}
\value{
BoundingPoly object
}
\description{
BoundingPoly Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A bounding polygon for the detected image annotation.
}

