% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{ColorInfo}
\alias{ColorInfo}
\title{ColorInfo Object}
\usage{
ColorInfo(pixelFraction = NULL, color = NULL, score = NULL)
}
\arguments{
\item{pixelFraction}{The fraction of pixels the color occupies in the image}

\item{color}{RGB components of the color}

\item{score}{Image-specific score for this color}
}
\value{
ColorInfo object
}
\description{
ColorInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Color information consists of RGB channels, score, and the fraction ofthe image that the color occupies in the image.
}
