% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{CropHint}
\alias{CropHint}
\title{CropHint Object}
\usage{
CropHint(boundingPoly = NULL, confidence = NULL,
  importanceFraction = NULL)
}
\arguments{
\item{boundingPoly}{The bounding polygon for the crop region}

\item{confidence}{Confidence of this being a salient region}

\item{importanceFraction}{Fraction of importance of this salient region with respect to the original}
}
\value{
CropHint object
}
\description{
CropHint Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Single crop hint that is used to generate a new crop when serving an image.
}
