% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{CropHintsAnnotation}
\alias{CropHintsAnnotation}
\title{CropHintsAnnotation Object}
\usage{
CropHintsAnnotation(cropHints = NULL)
}
\arguments{
\item{cropHints}{Crop hint results}
}
\value{
CropHintsAnnotation object
}
\description{
CropHintsAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Set of crop hints that are used to generate new crops when serving images.
}
