% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{CropHintsParams}
\alias{CropHintsParams}
\title{CropHintsParams Object}
\usage{
CropHintsParams(aspectRatios = NULL)
}
\arguments{
\item{aspectRatios}{Aspect ratios in floats, representing the ratio of the width to the height}
}
\value{
CropHintsParams object
}
\description{
CropHintsParams Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Parameters for crop hints annotation request.
}
