% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{DetectedBreak}
\alias{DetectedBreak}
\title{DetectedBreak Object}
\usage{
DetectedBreak(type = NULL, isPrefix = NULL)
}
\arguments{
\item{type}{Detected break type}

\item{isPrefix}{True if break prepends the element}
}
\value{
DetectedBreak object
}
\description{
DetectedBreak Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Detected start or end of a structural component.
}
