% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{DetectedLanguage}
\alias{DetectedLanguage}
\title{DetectedLanguage Object}
\usage{
DetectedLanguage(languageCode = NULL, confidence = NULL)
}
\arguments{
\item{languageCode}{The BCP-47 language code, such as 'en-US' or 'sr-Latn'}

\item{confidence}{Confidence of detected language}
}
\value{
DetectedLanguage object
}
\description{
DetectedLanguage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Detected language for a structural component.
}
