% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{DominantColorsAnnotation}
\alias{DominantColorsAnnotation}
\title{DominantColorsAnnotation Object}
\usage{
DominantColorsAnnotation(colors = NULL)
}
\arguments{
\item{colors}{RGB color values, with their score and pixel fraction}
}
\value{
DominantColorsAnnotation object
}
\description{
DominantColorsAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Set of dominant colors and their corresponding scores.
}

