% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{EntityAnnotation}
\alias{EntityAnnotation}
\title{EntityAnnotation Object}
\usage{
EntityAnnotation(mid = NULL, description = NULL, topicality = NULL,
  locale = NULL, properties = NULL, score = NULL, boundingPoly = NULL,
  locations = NULL, confidence = NULL)
}
\arguments{
\item{mid}{Opaque entity ID}

\item{description}{Entity textual description, expressed in its <code>locale</code> language}

\item{topicality}{The relevancy of the ICA (Image Content Annotation) label to the}

\item{locale}{The language code for the locale in which the entity textual}

\item{properties}{Some entities can have additional optional <code>Property</code> fields}

\item{score}{Overall score of the result}

\item{boundingPoly}{Image region to which this entity belongs}

\item{locations}{The location information for the detected entity}

\item{confidence}{The accuracy of the entity detection in an image}
}
\value{
EntityAnnotation object
}
\description{
EntityAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Set of detected entity features.
}

