% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{FaceAnnotation}
\alias{FaceAnnotation}
\title{FaceAnnotation Object}
\usage{
FaceAnnotation(tiltAngle = NULL, underExposedLikelihood = NULL,
  fdBoundingPoly = NULL, landmarkingConfidence = NULL,
  joyLikelihood = NULL, detectionConfidence = NULL,
  surpriseLikelihood = NULL, angerLikelihood = NULL,
  headwearLikelihood = NULL, panAngle = NULL, boundingPoly = NULL,
  landmarks = NULL, blurredLikelihood = NULL, rollAngle = NULL,
  sorrowLikelihood = NULL)
}
\arguments{
\item{tiltAngle}{Pitch angle}

\item{underExposedLikelihood}{Under-exposed likelihood}

\item{fdBoundingPoly}{This bounding polygon is tighter than the previous}

\item{landmarkingConfidence}{Face landmarking confidence}

\item{joyLikelihood}{Joy likelihood}

\item{detectionConfidence}{Detection confidence}

\item{surpriseLikelihood}{Surprise likelihood}

\item{angerLikelihood}{Anger likelihood}

\item{headwearLikelihood}{Headwear likelihood}

\item{panAngle}{Yaw angle}

\item{boundingPoly}{The bounding polygon around the face}

\item{landmarks}{Detected face landmarks}

\item{blurredLikelihood}{Blurred likelihood}

\item{rollAngle}{Roll angle}

\item{sorrowLikelihood}{Sorrow likelihood}
}
\value{
FaceAnnotation object
}
\description{
FaceAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A face annotation object contains the results of face detection.
}

