% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{Feature}
\alias{Feature}
\title{Feature Object}
\usage{
Feature(type = NULL, maxResults = NULL)
}
\arguments{
\item{type}{The feature type}

\item{maxResults}{Maximum number of results of this type}
}
\value{
Feature object
}
\description{
Feature Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Users describe the type of Google Cloud Vision API tasks to perform overimages by using *Feature*s. Each Feature indicates a type of imagedetection task to perform. Features encode the Cloud Vision APIvertical to operate on and the number of top-scoring results to return.
}
