% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{ImageContext}
\alias{ImageContext}
\title{ImageContext Object}
\usage{
ImageContext(languageHints = NULL, latLongRect = NULL,
  cropHintsParams = NULL)
}
\arguments{
\item{languageHints}{List of languages to use for TEXT_DETECTION}

\item{latLongRect}{lat/long rectangle that specifies the location of the image}

\item{cropHintsParams}{Parameters for crop hints annotation request}
}
\value{
ImageContext object
}
\description{
ImageContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Image context and/or feature-specific parameters.
}
