% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{ImageProperties}
\alias{ImageProperties}
\title{ImageProperties Object}
\usage{
ImageProperties(dominantColors = NULL)
}
\arguments{
\item{dominantColors}{If present, dominant colors completed successfully}
}
\value{
ImageProperties object
}
\description{
ImageProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Stores image properties (e.g. dominant colors).
}

