% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_functions.R
\name{images.annotate}
\alias{images.annotate}
\title{Run image detection and annotation for a batch of images.}
\usage{
images.annotate(BatchAnnotateImagesRequest)
}
\arguments{
\item{BatchAnnotateImagesRequest}{The \link{BatchAnnotateImagesRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/vision/}{Google Documentation}

Other BatchAnnotateImagesRequest functions: \code{\link{BatchAnnotateImagesRequest}}
}
