% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{ImageSource}
\alias{ImageSource}
\title{ImageSource Object}
\usage{
ImageSource(gcsImageUri = NULL, imageUri = NULL)
}
\arguments{
\item{gcsImageUri}{NOTE: For new code `image_uri` below is preferred}

\item{imageUri}{Image URI which supports:}
}
\value{
ImageSource object
}
\description{
ImageSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
External image source (Google Cloud Storage image location).
}
