% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{LatLongRect}
\alias{LatLongRect}
\title{LatLongRect Object}
\usage{
LatLongRect(minLatLng = NULL, maxLatLng = NULL)
}
\arguments{
\item{minLatLng}{Min lat/long pair}

\item{maxLatLng}{Max lat/long pair}
}
\value{
LatLongRect object
}
\description{
LatLongRect Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Rectangle determined by min and max `LatLng` pairs.
}
