% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{LocationInfo}
\alias{LocationInfo}
\title{LocationInfo Object}
\usage{
LocationInfo(latLng = NULL)
}
\arguments{
\item{latLng}{lat/long location coordinates}
}
\value{
LocationInfo object
}
\description{
LocationInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Detected entity location information.
}
