% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{Page}
\alias{Page}
\title{Page Object}
\usage{
Page(height = NULL, width = NULL, blocks = NULL, property = NULL)
}
\arguments{
\item{height}{Page height in pixels}

\item{width}{Page width in pixels}

\item{blocks}{List of blocks of text, images etc on this page}

\item{property}{Additional information detected on the page}
}
\value{
Page object
}
\description{
Page Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Detected page from OCR.
}
