% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{Paragraph}
\alias{Paragraph}
\title{Paragraph Object}
\usage{
Paragraph(words = NULL, property = NULL, boundingBox = NULL)
}
\arguments{
\item{words}{List of words in this paragraph}

\item{property}{Additional information detected for the paragraph}

\item{boundingBox}{The bounding box for the paragraph}
}
\value{
Paragraph object
}
\description{
Paragraph Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Structural unit of text representing a number of words in certain order.
}
