% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{Position}
\alias{Position}
\title{Position Object}
\usage{
Position(y = NULL, x = NULL, z = NULL)
}
\arguments{
\item{y}{Y coordinate}

\item{x}{X coordinate}

\item{z}{Z coordinate (or depth)}
}
\value{
Position object
}
\description{
Position Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A 3D position in the image, used primarily for Face detection landmarks.A valid Position must have both x and y coordinates.The position coordinates are in the same scale as the original image.
}

