% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{SafeSearchAnnotation}
\alias{SafeSearchAnnotation}
\title{SafeSearchAnnotation Object}
\usage{
SafeSearchAnnotation(violence = NULL, adult = NULL, spoof = NULL,
  medical = NULL)
}
\arguments{
\item{violence}{Violence likelihood}

\item{adult}{Represents the adult content likelihood for the image}

\item{spoof}{Spoof likelihood}

\item{medical}{Likelihood that this is a medical image}
}
\value{
SafeSearchAnnotation object
}
\description{
SafeSearchAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Set of features pertaining to the image, computed by computer visionmethods over safe-search verticals (for example, adult, spoof, medical,violence).
}
