% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{Symbol}
\alias{Symbol}
\title{Symbol Object}
\usage{
Symbol(text = NULL, property = NULL, boundingBox = NULL)
}
\arguments{
\item{text}{The actual UTF-8 representation of the symbol}

\item{property}{Additional information detected for the symbol}

\item{boundingBox}{The bounding box for the symbol}
}
\value{
Symbol object
}
\description{
Symbol Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single symbol representation.
}
