% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{TextAnnotation}
\alias{TextAnnotation}
\title{TextAnnotation Object}
\usage{
TextAnnotation(pages = NULL, text = NULL)
}
\arguments{
\item{pages}{List of pages detected by OCR}

\item{text}{UTF-8 text detected on the pages}
}
\value{
TextAnnotation object
}
\description{
TextAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
TextAnnotation contains a structured representation of OCR extracted text.The hierarchy of an OCR extracted text structure is like this:    TextAnnotation -> Page -> Block -> Paragraph -> Word -> SymbolEach structural component, starting from Page, may further have their ownproperties. Properties describe detected languages, breaks etc.. Pleaserefer to the google.cloud.vision.v1.TextAnnotation.TextProperty messagedefinition below for more detail.
}
