% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{TextProperty}
\alias{TextProperty}
\title{TextProperty Object}
\usage{
TextProperty(detectedBreak = NULL, detectedLanguages = NULL)
}
\arguments{
\item{detectedBreak}{Detected start or end of a text segment}

\item{detectedLanguages}{A list of detected languages together with confidence}
}
\value{
TextProperty object
}
\description{
TextProperty Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional information detected on the structural component.
}
