% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{Vertex}
\alias{Vertex}
\title{Vertex Object}
\usage{
Vertex(y = NULL, x = NULL)
}
\arguments{
\item{y}{Y coordinate}

\item{x}{X coordinate}
}
\value{
Vertex object
}
\description{
Vertex Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A vertex represents a 2D point in the image.NOTE: the vertex coordinates are in the same scale as the original image.
}

