% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_functions.R
\docType{package}
\name{vision_googleAuthR}
\alias{vision_googleAuthR}
\alias{vision_googleAuthR-package}
\title{Google Cloud Vision API
Integrates Google Vision features, including image labeling, face, logo, and landmark detection, optical character recognition (OCR), and detection of explicit content, into applications.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:50:21
filename: /Users/mark/dev/R/autoGoogleAPI/googlevisionv1.auto/R/vision_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}
}

