% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{WebDetection}
\alias{WebDetection}
\title{WebDetection Object}
\usage{
WebDetection(fullMatchingImages = NULL, webEntities = NULL,
  pagesWithMatchingImages = NULL, partialMatchingImages = NULL)
}
\arguments{
\item{fullMatchingImages}{Fully matching images from the Internet}

\item{webEntities}{Deduced entities from similar images on the Internet}

\item{pagesWithMatchingImages}{Web pages containing the matching images from the Internet}

\item{partialMatchingImages}{Partial matching images from the Internet}
}
\value{
WebDetection object
}
\description{
WebDetection Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Relevant information for the image from the Internet.
}
