% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{WebEntity}
\alias{WebEntity}
\title{WebEntity Object}
\usage{
WebEntity(entityId = NULL, description = NULL, score = NULL)
}
\arguments{
\item{entityId}{Opaque entity ID}

\item{description}{Canonical description of the entity, in English}

\item{score}{Overall relevancy score for the entity}
}
\value{
WebEntity object
}
\description{
WebEntity Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Entity deduced from similar images on the Internet.
}
