% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{WebImage}
\alias{WebImage}
\title{WebImage Object}
\usage{
WebImage(score = NULL, url = NULL)
}
\arguments{
\item{score}{Overall relevancy score for the image}

\item{url}{The result image URL}
}
\value{
WebImage object
}
\description{
WebImage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata for online images.
}
