% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{WebPage}
\alias{WebPage}
\title{WebPage Object}
\usage{
WebPage(url = NULL, score = NULL)
}
\arguments{
\item{url}{The result web page URL}

\item{score}{Overall relevancy score for the web page}
}
\value{
WebPage object
}
\description{
WebPage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata for web pages.
}
