% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{Word}
\alias{Word}
\title{Word Object}
\usage{
Word(property = NULL, boundingBox = NULL, symbols = NULL)
}
\arguments{
\item{property}{Additional information detected for the word}

\item{boundingBox}{The bounding box for the word}

\item{symbols}{List of symbols in the word}
}
\value{
Word object
}
\description{
Word Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A word representation.
}
