% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{AnnotateImageResponse}
\alias{AnnotateImageResponse}
\title{AnnotateImageResponse Object}
\usage{
AnnotateImageResponse(labelAnnotations = NULL, landmarkAnnotations = NULL,
  safeSearchAnnotation = NULL, imagePropertiesAnnotation = NULL,
  textAnnotations = NULL, logoAnnotations = NULL, faceAnnotations = NULL,
  error = NULL)
}
\arguments{
\item{labelAnnotations}{If present, label detection completed successfully}

\item{landmarkAnnotations}{If present, landmark detection completed successfully}

\item{safeSearchAnnotation}{If present, safe-search annotation completed successfully}

\item{imagePropertiesAnnotation}{If present, image properties were extracted successfully}

\item{textAnnotations}{If present, text (OCR) detection completed successfully}

\item{logoAnnotations}{If present, logo detection completed successfully}

\item{faceAnnotations}{If present, face detection completed successfully}

\item{error}{If set, represents the error message for the operation}
}
\value{
AnnotateImageResponse object
}
\description{
AnnotateImageResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response to an image annotation request.
}

