% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{BatchAnnotateImagesRequest}
\alias{BatchAnnotateImagesRequest}
\title{BatchAnnotateImagesRequest Object}
\usage{
BatchAnnotateImagesRequest(requests = NULL)
}
\arguments{
\item{requests}{Individual image annotation requests for this batch}
}
\value{
BatchAnnotateImagesRequest object
}
\description{
BatchAnnotateImagesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Multiple image annotation requests are batched into a single service call.
}
\seealso{
Other BatchAnnotateImagesRequest functions: \code{\link{images.annotate}}
}

