% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{EntityAnnotation}
\alias{EntityAnnotation}
\title{EntityAnnotation Object}
\usage{
EntityAnnotation(boundingPoly = NULL, locale = NULL, topicality = NULL,
  description = NULL, properties = NULL, score = NULL, locations = NULL,
  mid = NULL, confidence = NULL)
}
\arguments{
\item{boundingPoly}{Image region to which this entity belongs}

\item{locale}{The language code for the locale in which the entity textual}

\item{topicality}{The relevancy of the ICA (Image Content Annotation) label to the}

\item{description}{Entity textual description, expressed in its `locale` language}

\item{properties}{Some entities may have optional user-supplied `Property` (name/value)}

\item{score}{Overall score of the result}

\item{locations}{The location information for the detected entity}

\item{mid}{Opaque entity ID}

\item{confidence}{The accuracy of the entity detection in an image}
}
\value{
EntityAnnotation object
}
\description{
EntityAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Set of detected entity features.
}
