% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{Image}
\alias{Image}
\title{Image Object}
\usage{
Image(source = NULL, content = NULL)
}
\arguments{
\item{source}{Google Cloud Storage image location}

\item{content}{Image content, represented as a stream of bytes}
}
\value{
Image object
}
\description{
Image Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Client image to perform Google Cloud Vision API tasks over.
}

