% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{ImageContext}
\alias{ImageContext}
\title{ImageContext Object}
\usage{
ImageContext(latLongRect = NULL, languageHints = NULL)
}
\arguments{
\item{latLongRect}{Lat/long rectangle that specifies the location of the image}

\item{languageHints}{List of languages to use for TEXT_DETECTION}
}
\value{
ImageContext object
}
\description{
ImageContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Image context.
}

