% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{ImageSource}
\alias{ImageSource}
\title{Google Cloud Vision API Objects 
Integrates Google Vision features, including image labeling, face, logo, and landmark detection, optical character recognition (OCR), and detection of explicit content, into applications.}
\usage{
ImageSource(gcsImageUri = NULL)
}
\arguments{
\item{gcsImageUri}{Google Cloud Storage image URI}
}
\value{
ImageSource object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:50:21
filename: /Users/mark/dev/R/autoGoogleAPI/googlevisionv1.auto/R/vision_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ImageSource Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
External image source (Google Cloud Storage image location).
}

