% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{Property}
\alias{Property}
\title{Property Object}
\usage{
Property(value = NULL, name = NULL)
}
\arguments{
\item{value}{Value of the property}

\item{name}{Name of the property}
}
\value{
Property object
}
\description{
Property Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Arbitrary name/value pair.
}

