% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_objects.R
\name{SafeSearchAnnotation}
\alias{SafeSearchAnnotation}
\title{SafeSearchAnnotation Object}
\usage{
SafeSearchAnnotation(medical = NULL, spoof = NULL, violence = NULL,
  adult = NULL)
}
\arguments{
\item{medical}{Likelihood this is a medical image}

\item{spoof}{Spoof likelihood}

\item{violence}{Violence likelihood}

\item{adult}{Represents the adult contents likelihood for the image}
}
\value{
SafeSearchAnnotation object
}
\description{
SafeSearchAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Set of features pertaining to the image, computed by various computer visionmethods over safe-search verticals (for example, adult, spoof, medical,violence).
}

